requirements.txt
pyperclip
cryptography
pyyaml
click
zxcvbn